+++ libs/safe/src/lib/survey/init.ts
@@ -1,7 +1,5 @@
 // This is needed for compilation of some packages with strict option enabled.
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference
-/// <reference path="../../typings/surveyjs-widgets/index.d.ts" />
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference
 /// <reference path="../../typings/extract-files/index.d.ts" />
 
 import { Apollo } from 'apollo-angular';
@@ -23,6 +21,7 @@ import * as ReferenceDataProperties from './global-properties/reference-data';
 import * as TooltipProperty from './global-properties/tooltip';
 import { initLocalization } from './localization';
 import { Dialog } from '@angular/cdk/dialog';
+import { NgZone } from '@angular/core';
 
 /**
  * Executes all init methods of custom SurveyJS.
@@ -36,6 +35,7 @@ import { Dialog } from '@angular/cdk/dialog';
  * @param environment injected environment
  * @param referenceDataService Reference data service
  * @param containsCustomQuestions If survey contains custom questions or not
+ * @param ngZone Angular Service to execute code inside Angular environment
  */
 export const initCustomSurvey = (
   Survey: any,
@@ -46,7 +46,8 @@ export const initCustomSurvey = (
   authService: SafeAuthService,
   environment: any,
   referenceDataService: SafeReferenceDataService,
-  containsCustomQuestions: boolean
+  containsCustomQuestions: boolean,
+  ngZone: NgZone
 ): void => {
   // If the survey created does not contain custom questions, we destroy previously set custom questions if so
   if (!containsCustomQuestions) {
@@ -61,8 +62,22 @@ export const initCustomSurvey = (
   if (containsCustomQuestions) {
     CommentWidget.init(Survey);
     // load components (same as widgets, but with less configuration options)
-    ResourceComponent.init(Survey, domService, apollo, dialog, formBuilder);
-    ResourcesComponent.init(Survey, domService, apollo, dialog, formBuilder);
+    ResourceComponent.init(
+      Survey,
+      domService,
+      apollo,
+      dialog,
+      formBuilder,
+      ngZone
+    );
+    ResourcesComponent.init(
+      Survey,
+      domService,
+      apollo,
+      dialog,
+      formBuilder,
+      ngZone
+    );
     OwnerComponent.init(Survey, domService, apollo);
     UsersComponent.init(Survey, domService, apollo);
   }